+++ test/bad_dtls_test.c
@@ -335,7 +335,7 @@ static int send_record(BIO *rbio, unsigned char type, uint64_t seqnr,
             || !TEST_ptr(enc_ctx = EVP_CIPHER_CTX_new())
             || !TEST_true(EVP_CipherInit_ex(enc_ctx, EVP_aes_128_cbc(), NULL,
                                             enc_key, iv, 1))
-            || !TEST_int_ge(EVP_Cipher(enc_ctx, enc, enc, len), 0))
+            || !TEST_int_ge(EVP_Cipher(enc_ctx, enc, enc, (unsigned int)len), 0))
         goto end;
 
     /* Finally write header (from fragmented variables), IV and encrypted record */
@@ -348,7 +348,7 @@ static int send_record(BIO *rbio, unsigned char type, uint64_t seqnr,
     BIO_write(rbio, lenbytes, 2);
 
     BIO_write(rbio, iv, sizeof(iv));
-    BIO_write(rbio, enc, len);
+    BIO_write(rbio, enc, (int)len);
     ret = 1;
  end:
     EVP_MAC_free(hmac);
