+++ test/danetest.c
@@ -115,7 +115,7 @@ static STACK_OF(X509) *load_chain(BIO *fp, int nelem)
             const unsigned char *p = data;
 
             if (!TEST_ptr(cert = d(0, &p, len))
-                    || !TEST_long_eq(p - data, len)) {
+                    || !TEST_long_eq((long)(p - data), len)) {
                 TEST_info("Certificate parsing error");
                 goto err;
             }
@@ -155,7 +155,7 @@ static char *read_to_eol(BIO *f)
     if (BIO_gets(f, buf, sizeof(buf)) <= 0)
         return NULL;
 
-    n = strlen(buf);
+    n = (int)strlen(buf);
     if (buf[n - 1] != '\n') {
         if (n + 1 == sizeof(buf))
             TEST_error("input too long");
